Fix generate tarball flags for both subcommands
